requirements.txt
amqp==5.0.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bleach==4.1.0
Brotli==1.0.9
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.2
cryptography==36.0.1
cssselect2==0.4.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-allauth==0.47.0
django-appconf==1.0.5
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==21.1
django-bootstrap3-multidatepicker==0.1.5
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-filter==21.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-money==2.1
django-mptt==0.13.4
django-phonenumber-field==6.0.0
django-select2==7.9.0
django-webpush==0.3.3
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
entrypoints==0.3
fonttools==4.28.4
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
http-ece==1.1.0
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonfield==3.1.0
jsonschema==4.3.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
kombu==5.2.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mplcursors==0.5.1
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.39
pickleshare==0.7.5
Pillow==8.4.0
playsound==1.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py-moneyed==1.2
py-vapid==1.8.2
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
pyphen==0.11.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywebpush==1.9.4
pyzmq==22.3.0
redis==4.0.2
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
Send2Trash==1.8.0
sentry-sdk==1.5.1
six==1.16.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
swapper==1.3.0
terminado==0.12.1
testpath==0.5.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
twilio==7.4.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
weasyprint==53.4
webencodings==0.5.1
wrapt==1.13.3
zipp==3.6.0
zopfli==0.1.9
